+++ doc/quick-start/code/index.mjs
@@ -3,10 +3,9 @@
 
 import { default as CONFIG } from './config-framework.mjs';
 
-import { Host, KeepRunning } from '@this/host';
+import { BaseSystem, Host, KeepRunning } from '@this/host';
 import { Loggy } from '@this/loggy';
 import { WebappRoot } from '@this/webapp-core';
-import { BaseSystem } from '@this/webapp-util';
 
 
 /**