+++ src/webapp-util/index.js
@@ -2,6 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from '#x/BaseFileService';
-export * from '#x/BaseSystem';
 export * from '#x/Rotator';
 export * from '#x/Saver';